move utils to fetch.server (fixing azure)
