+++ packages/vue-app/template/mixins/fetch.server.js
@@ -1,12 +1,16 @@
 import Vue from 'vue'
-import { watchDiff, hasFetch, normalizeError, getDataDiff, addLifecycleHook } from '../utils'
+import { hasFetch, normalizeError, getDataDiff, addLifecycleHook } from '../utils'
 
 async function serverPrefetch() {
   if (!this._fetchOnServer) {
     return
   }
 
-  const diff = watchDiff(this)
+  // Watch data mutations during fetch call
+  const [data, diff] = watchDiff(vm._data)
+  vm._data = data
+
+  // Call and await on $fetch
   try {
     await this.$options.fetch.call(this)
   } catch (err) {
@@ -22,6 +26,54 @@ async function serverPrefetch() {
 
   // Call asyncData & add to ssrContext for window.__NUXT__.asyncData
   this.$ssrContext.nuxt.data.push(this.$fetchState.error ? { _error: this.$fetchState.error } : diff)
+}
+
+function isObject(obj) {
+  return typeof obj === 'object' && obj !== null && !Array.isArray(obj)
+}
+
+function isPrimitive(val) {
+  return val !== Object(val);
+}
+
+function watchDiff(obj) {
+  const diff = {}
+
+  const watchedProps = new Set()
+
+  const proxy = new Proxy(obj, {
+    get(_, prop) {
+      const value = obj[prop]
+
+      // Skip already watched props and hidden props like __ob__
+      if (prop[0] == '_' || watchedProps.has(prop)) {
+        return value
+      }
+
+      // Lazy deep-watch
+      if (isObject(value)) {
+        const [_value, _diff] = _watchDiff(value)
+        watchedProps.add(prop)
+        obj[prop] = _value
+        diff[prop] = _diff
+        return _value
+      } else if (value !== undefined && !isPrimitive(value)) {
+        // We can't predict mutations so always mark as dirty
+        diff[prop] = value
+      }
+
+      return value
+    },
+    set(_, prop, val) {
+      obj[prop] = val
+      if (prop[0] !== '_') {
+        diff[prop] = val
+      }
+      return true
+    }
+  })
+
+  return [proxy, diff]
 }
 
 export default {
